requirements.txt
python-telegram-bot>=20.7
fal-client>=0.7.0
python-dotenv>=1.0.0
aiohttp>=3.9.1
Pillow>=10.0.0
requests>=2.31.0 